+++ src/containers/InteractionMonitoring/index.js
@@ -27,11 +27,13 @@ import {
   selectInteractionMonitoringSelected,
   userHasViewAllMonitoredCallsPermission,
   userHasMonitorAllCallsPermission,
-  selectExtensions
+  selectExtensions,
 } from '../../redux/modules/reporting/interactionMonitoring/selectors';
 import {
   selectSupervisorToolbarSilentMonitoringInteractionId,
-  selectSupervisorToolbarSilentMonitoringStatus
+  selectSupervisorToolbarSilentMonitoringStatus,
+  selectCanSilentMonitor,
+  selectLoadingUserStatus
 } from '../../redux/modules/supervisorToolbar/selectors';
 import { getCurrentAgentId } from '../../redux/modules/userData/selectors';
 import { getHelpLink } from '../EntityTable/selectors';
@@ -40,7 +42,8 @@ import { entitiesMetaData } from '../../redux/modules/entities/metaData';
 import Layout from './layout';
 import { isInIframe } from 'serenova-js-utils/browser';
 
-export const mapStateToProps = (state, props) => ({
+export const mapStateToProps = (state, props) => {
+  return {
   pageTitle: entitiesMetaData['interactionMonitoring'] ? entitiesMetaData['interactionMonitoring'].pageTitle : '',
   pageHelpLink: getHelpLink(state),
   getCurrentAgentId: getCurrentAgentId(state),
@@ -60,8 +63,10 @@ export const mapStateToProps = (state, props) => ({
   monitoringStatus: selectSupervisorToolbarSilentMonitoringStatus(state, props),
   userHasViewAllMonitoredCallsPermission: userHasViewAllMonitoredCallsPermission(state),
   userHasMonitorAllCallsPermission: userHasMonitorAllCallsPermission(state),
-  insideIframe: !isInIframe()
-});
+  insideIframe: !isInIframe(),
+  canSilentMonitor: selectCanSilentMonitor(state),
+  loadingUserStatus: selectLoadingUserStatus(state)
+}};
 
 export const actions = {
   toggleTimeFormat,